split proxy_cli
