+++ src/lib.rs
@@ -5,5 +5,6 @@ pub mod file;
 pub mod fuelup_cli;
 pub mod ops;
 pub mod path;
+pub mod proxy_cli;
 pub mod settings;
 pub mod toolchain;
